requirements.txt
selenium
requests
transformers
Pillow
customtkinter
fpdf 